refactor(dev-dasboard): rename & update components
