+++ ui/apps/dev-dashboard/src/components/onboarding/index.ts
@@ -1,0 +1,4 @@
+export * from './download-cli-tool';
+export * from './generate-message';
+export * from './key-confirmation';
+export * from './terms-and-conditions';
